pyproject.toml
[project]
name = "mcp-remote-execution"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "e2b-code-interpreter>=2.0.0",
    "langchain-openai>=0.3.34",
    "mcp-use>=1.3.11",
    "mcp[cli]>=1.15.0",
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
]
